Publicly expose fragment_helpers module
